APIs/Biquery/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiosqlite==0.16.1
Async-OAuthlib==0.0.9
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
chevron==0.14.0
httptools==0.2.0
idna==2.10
iso8601==0.1.14
multidict==5.1.0
oauthlib==3.1.1
PyMySQL==0.9.3
pypika-tortoise==0.1.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
sanic==21.6.0
Sanic-Cors==1.0.0
sanic-plugin-toolkit==1.0.1
sanic-routing==0.7.0
sanic-session==0.8.0
tortoise-orm==0.17.5
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvloop==0.15.2
websockets==9.1
yarl==1.6.3
